Track the Blocks & Transactions in an embedded db
* We need to maintain our own db so we can query the chain more easily
* (for instance to get a list of transactions for a given contract)

+++ src/main/java/com/jpmorgan/ib/caonpd/ethereum/enterprise/controller/BlockController.java
@@ -29,7 +29,7 @@ public class BlockController extends BaseController {
     public ResponseEntity<APIResponse> getBlock(
             @JsonBodyParam(required=false) String id,
             @JsonBodyParam(required=false) String hash,
-            @JsonBodyParam(required=false) Integer number,
+            @JsonBodyParam(required=false) Long number,
             @JsonBodyParam(required=false) String tag) throws APIException {
 
         if (id == null && hash != null) {
